refactor dpm exception ai route
